requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coincurve==17.0.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
eciespy==0.3.13
eth-hash==0.5.1
eth-keys==0.4.0
eth-typing==3.2.0
eth-utils==2.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
toolz==0.12.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
